Refactor Story Library
