+++ database/data_loader.py
@@ -21,9 +21,9 @@ sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 from botocore.exceptions import ClientError
 
-from eidolon.dynamo import dynamo  # noqa: E402
+from eidolon.dynamo import dynamo
 from eidolon.dynamo import TableName
-from eidolon.validation import validate_character_name  # noqa: E402
+from eidolon.validation import validate_character_name
 
 
 def load_json(file_path):
