+++ deployment/stacks/client_stack.py
@@ -1,6 +1,6 @@
 """Client stack for portal deployment with CloudFront and CodeBuild."""
 
-from aws_cdk import CfnOutput, Duration, RemovalPolicy, Stack
+from aws_cdk import CfnOutput, Duration, RemovalPolicy, Stack, Tags
 from aws_cdk import aws_certificatemanager as acm
 from aws_cdk import aws_cloudfront as cloudfront
 from aws_cdk import aws_cloudfront_origins as origins
@@ -74,6 +74,9 @@ class ClientStack(Stack):
         self.bucket_exists = bucket_exists
 
         super().__init__(scope, stack_id, description="Portal S3 bucket, CloudFront CDN, and CodeBuild deployment", **kwargs)
+
+        # Add system tag to all resources in this stack
+        Tags.of(self).add("System", "Eidolon")
 
         # Create portal S3 bucket
         self.portal_bucket = self._create_portal_bucket()
