+++ deployment/stacks/codebuild_stack.py
@@ -1,6 +1,6 @@
 """CodeBuild stack for Eidolon Engine Lambda builds."""
 
-from aws_cdk import CfnOutput, RemovalPolicy, Stack
+from aws_cdk import CfnOutput, RemovalPolicy, Stack, Tags
 from aws_cdk import aws_codebuild as codebuild
 from aws_cdk import aws_iam as iam
 from aws_cdk import aws_s3 as s3
@@ -41,6 +41,8 @@ class CodeBuildStack(Stack):
         self.github_repo = github_repo
         self.github_branch = github_branch
         super().__init__(scope, stack_id, **kwargs)
+        # Apply system tag to all resources in this stack
+        Tags.of(self).add("System", "Eidolon")
 
         # Import existing bucket or create new one with fixed logical ID
         if bucket_exists:
