--- eidolon/story.py
@@ -1,1379 +1,0 @@
-"""
-Story management utilities for Lambda functions.
-
-Provides common functions for story operations including starting, abandoning,
-and managing story segments.
-"""
-
-import time
-from datetime import datetime, timezone
-from decimal import Decimal
-
-from botocore.exceptions import ClientError
-from uuid_extension import uuid7
-
-from eidolon.character_segment import update_character_active_segment
-from eidolon.character_story import get_story_history
-from eidolon.dynamo import TableName, dynamo
-from eidolon.environment import SEGMENT_QUEUE_URL
-from eidolon.logger import logger
-from eidolon.player import verify_character_ownership
-from eidolon.schema import normalize_segment_definition
-from eidolon.segment_core import get_segment_definition
-from eidolon.segment_history import record_segment_history
-from eidolon.segment_polling import delete_active_segment
-from eidolon.segment_processing import determine_next_segment
-from eidolon.segment_state import create_next_active_segment
-from eidolon.sqs import send_message
-from eidolon.time_utils import now_iso, future_iso, now_unix, future_unix, from_unix
-
-
-def get_active_story_segment(character_id: str) -> dict:
-    """
-    Get the active story segment for a character.
-
-    Args:
-        character_id: Character UUID
-
-    Returns:
-        Active segment data dict
-
-    Raises:
-        ValueError: If no active story found for character
-        RuntimeError: If database query fails
-    """
-    if not character_id:
-        raise ValueError("Character ID cannot be empty")
-
-    try:
-        items = dynamo.query(
-            TableName.ACTIVE_SEGMENTS,
-            IndexName="CharacterID-index",
-            KeyConditionExpression="CharacterID = :cid",
-            FilterExpression="#status = :status",
-            ExpressionAttributeNames={"#status": "Status"},
-            ExpressionAttributeValues={":cid": character_id, ":status": "active"},
-        )
-    except ClientError as err:
-        logger.error(f"Failed to query active segments for {character_id} Error: {err}", exc_info=True)
-        raise RuntimeError from err
-
-    if not items:
-        raise ValueError(f"No active story found for character {character_id}")
-
-    return items[0]
-
-
-def mark_segment_as_abandoned(active_segment_id: str) -> None:
-    """
-    Mark an active segment as abandoned.
-
-    Args:
-        active_segment_id: Active segment UUID
-
-    Raises:
-        ValueError: If active_segment_id is empty
-        RuntimeError: If database update fails
-    """
-    if not active_segment_id:
-        raise ValueError("Active segment ID cannot be empty")
-
-    try:
-        dynamo.update_item(
-            TableName.ACTIVE_SEGMENTS,
-            Key={"ActiveSegmentID": active_segment_id},
-            UpdateExpression="SET #status = :status",
-            ExpressionAttributeNames={"#status": "Status"},
-            ExpressionAttributeValues={":status": "abandoned"},
-        )
-    except ClientError as err:
-        logger.error(f"Failed to mark segment as abandoned for {active_segment_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to mark segment as abandoned: {err}") from err
-
-    logger.info(f"Marked segment as abandoned for {active_segment_id}")
-
-
-def record_story_abandonment(character_id: str, story_instance_id: str) -> None:
-    """
-    Update history to record story abandonment.
-
-    Args:
-        character_id: Character UUID
-        story_instance_id: Story instance UUID (UUIDv7)
-
-    Raises:
-        ValueError: If character_id or story_instance_id is empty
-        RuntimeError: If database operations fail
-    """
-    if not character_id:
-        raise ValueError("Character ID cannot be empty")
-    if not story_instance_id:
-        raise ValueError("Story instance ID cannot be empty")
-
-    try:
-        dynamo.update_item(
-            TableName.STORY_HISTORY,
-            Key={"CharacterID": character_id, "StoryInstanceID": story_instance_id},
-            UpdateExpression="SET FinishedAt = :finished, FinalOutcome = :outcome",
-            ExpressionAttributeValues={
-                ":finished": now_iso(),
-                ":outcome": "abandoned",
-            },
-        )
-    except ClientError as err:
-        logger.error(f"Failed to update story history for {character_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to update story history: {err}") from err
-
-    logger.info(f"Updated story history with abandonment for {character_id}")
-
-
-def add_story_to_abandoned_list(character_id: str, story_id: str) -> None:
-    """
-    Add a story to the character's AbandonedStories list.
-
-    Uses DynamoDB's ADD operation which automatically handles duplicates,
-    ensuring each story ID appears only once in the set.
-
-    Args:
-        character_id: Character UUID
-        story_id: Story UUID to add to abandoned list
-
-    Raises:
-        ValueError: If character_id or story_id is empty
-        RuntimeError: If database update fails
-    """
-    if not character_id:
-        raise ValueError("Character ID cannot be empty")
-    if not story_id:
-        raise ValueError("Story ID cannot be empty")
-
-    try:
-        dynamo.update_item(
-            TableName.CHARACTERS,
-            Key={"CharacterID": character_id},
-            UpdateExpression="ADD AbandonedStories :story",
-            ExpressionAttributeValues={":story": {story_id}},
-        )
-    except ClientError as err:
-        logger.error(f"Failed to add story to abandoned list for {character_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to add story to abandoned list: {err}") from err
-
-    logger.info(f"Added story to abandoned list for {character_id}")
-
-
-def get_active_story_segment_with_player_check(character_id: str, player_id: str) -> dict:
-    """
-    Get the active story segment for a character with player ID verification.
-
-    Args:
-        character_id: Character UUID
-        player_id: Player ID for verification
-
-    Returns:
-        Active segment data dict
-
-    Raises:
-        ValueError: If no active story found for character
-        RuntimeError: If database query fails
-    """
-    if not character_id:
-        raise ValueError("Character ID cannot be empty")
-    if not player_id:
-        raise ValueError("Player ID cannot be empty")
-
-    try:
-        items = dynamo.query(
-            TableName.ACTIVE_SEGMENTS,
-            IndexName="CharacterID-index",
-            KeyConditionExpression="CharacterID = :cid",
-            FilterExpression="PlayerID = :pid AND #status = :status",
-            ExpressionAttributeNames={"#status": "Status"},
-            ExpressionAttributeValues={
-                ":cid": character_id,
-                ":pid": player_id,
-                ":status": "active",
-            },
-        )
-    except ClientError as err:
-        logger.error(f"Failed to query active segments for {character_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to query active segments: {err}") from err
-
-    if not items:
-        raise ValueError("No active story found")
-
-    return items[0]
-
-
-def get_story_segment(story_id: str, segment_id: str) -> dict:
-    """
-    Get a specific segment from the SEGMENTS table.
-
-    Args:
-        story_id: Story UUID
-        segment_id: Segment UUID
-
-    Returns:
-        Segment data dict
-
-    Raises:
-        ValueError: If segment not found
-        RuntimeError: If database query fails
-    """
-    if not story_id:
-        raise ValueError("Story ID cannot be empty")
-    if not segment_id:
-        raise ValueError("Segment ID cannot be empty")
-
-    try:
-        segment = dynamo.get_item(TableName.SEGMENTS, {"StoryID": story_id, "SegmentID": segment_id})
-        if not segment:
-            raise ValueError("Segment not found")
-        return normalize_segment_definition(segment)
-    except ClientError as err:
-        logger.error(f"Failed to get segment for {segment_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to get segment: {err}") from err
-
-
-def get_completed_segment_for_character(character_id: str, player_id: str, segment_id: str) -> dict:
-    """
-    Get a completed segment for a character with player ID verification.
-
-    Args:
-        character_id: Character UUID
-        player_id: Player ID for verification
-        segment_id: Segment UUID to find
-
-    Returns:
-        Active segment data dict
-
-    Raises:
-        ValueError: If segment not found or not completed
-        RuntimeError: If database query fails
-    """
-    if not character_id:
-        raise ValueError("Character ID cannot be empty")
-    if not player_id:
-        raise ValueError("Player ID cannot be empty")
-    if not segment_id:
-        raise ValueError("Segment ID cannot be empty")
-
-    try:
-        items = dynamo.query(
-            TableName.ACTIVE_SEGMENTS,
-            IndexName="CharacterID-index",
-            KeyConditionExpression="CharacterID = :cid",
-            # Accept both processed and completed to support pre-completion outcomes
-            FilterExpression="PlayerID = :pid AND SegmentID = :sid AND (#status = :completed OR ProcessingStatus = :processed)",
-            ExpressionAttributeNames={"#status": "Status"},
-            ExpressionAttributeValues={
-                ":cid": character_id,
-                ":pid": player_id,
-                ":sid": segment_id,
-                ":completed": "completed",
-                ":processed": "processed",
-            },
-        )
-    except ClientError as err:
-        logger.error(f"Failed to query active segments for {character_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to query segments: {err}") from err
-
-    if not items:
-        # Fallback: check segment_history for the last record for this character/segment
-        try:
-            history_items = dynamo.query(
-                TableName.SEGMENT_HISTORY,
-                KeyConditionExpression="CharacterID = :cid",
-                ExpressionAttributeValues={":cid": character_id},
-            )
-        except ClientError as err:
-            logger.error(
-                f"Failed to query segment history for {character_id} Error: {err}",
-                exc_info=True,
-            )
-            raise RuntimeError(f"Failed to query segments: {err}") from err
-
-        if not history_items:
-            raise ValueError("Completed segment not found")
-
-        # Filter to matching SegmentID; choose the latest by CompletedAt (fallback EndTime)
-        candidates = [it for it in history_items if it.get("SegmentID") == segment_id]
-        if not candidates:
-            raise ValueError("Completed segment not found")
-
-        def sort_key(it: dict):
-            completed = it.get("CompletedAt") or ""
-            end_time = it.get("EndTime") or 0
-            return (1, completed) if completed else (0, end_time)
-
-        return max(candidates, key=sort_key)
-
-    active_segment = items[0]
-
-    # If only processed (outcome known pre-completion), allow
-    if active_segment.get("Status") != "completed" and active_segment.get("ProcessingStatus") != "processed":
-        raise ValueError("Segment not yet completed")
-
-    return active_segment
-
-
-def get_story_cooldown(character_id: str, story_id: str, story_type: str):
-    """
-    Calculate cooldown remaining for a story based on its type and last completion.
-
-    Args:
-        character_id: Character UUID
-        story_id: Story UUID
-        story_type: Type of story (one-time, daily, repeatable)
-
-    Returns:
-        Seconds remaining on cooldown, 0 if playable, -1 if permanently unavailable
-    """
-    if story_type == "repeatable":
-        return 0
-
-    try:
-        history = get_story_history(character_id, story_id)
-
-        if not history:
-            return 0  # Never played
-
-        # Check if story was completed or abandoned
-        if not history.get("FinishedAt"):
-            return 0  # Abandoned stories can be retried
-
-        if story_type == "one-time":
-            # Check if it was completed successfully
-            outcome = history.get("FinalOutcome", "")
-            if outcome in ["normal", "exceptional", "minimal"]:
-                return -1  # Permanently unavailable
-            return 0  # Failed/died, can retry
-
-        if story_type == "daily":
-            # Calculate time until midnight UTC
-            finished_at = datetime.fromisoformat(history.get("FinishedAt", "").replace("Z", "+00:00"))
-            now = datetime.now(timezone.utc)
-
-            # Check if completion was today
-            if finished_at.date() == now.date():
-                # Calculate seconds until midnight
-                midnight = now.replace(hour=0, minute=0, second=0, microsecond=0)
-                midnight = midnight.replace(day=midnight.day + 1)
-                return int((midnight - now).total_seconds())
-
-            return 0  # Completed on a previous day
-
-    except Exception as err:
-        logger.error(f"Error checking story cooldown Error: {err}")
-        return 0
-
-
-def check_story_prerequisites(character: dict, prerequisites: dict) -> bool:
-    """
-    Check if character meets story prerequisites.
-
-    Args:
-        character: Character data
-        prerequisites: Story prerequisite requirements
-
-    Returns:
-        True if all prerequisites are met
-    """
-    # Check minimum skills
-    min_skills = prerequisites.get("minSkills", {})
-    character_skills = character.get("Skills", {})
-
-    for skill, min_value in min_skills.items():
-        if character_skills.get(skill, 0) < min_value:
-            return False
-
-    # Check required items
-    required_items = prerequisites.get("requiredItems", [])
-    if required_items:
-        inventory = character.get("Inventory", {})
-        inventory_items = list(inventory.values())
-        for item_id in required_items:
-            if item_id not in inventory_items:
-                return False
-
-    return True
-
-
-def validate_story_available(character: dict, story_id: str) -> None:
-    """
-    Validate that the story is available to the character.
-
-    Args:
-        character: Character data
-        story_id: Story UUID to start
-
-    Raises:
-        ValueError: If story not available to character
-    """
-    available_stories = character.get("AvailableStories", [])
-    if story_id not in available_stories:
-        raise ValueError("Story not available")
-
-
-def get_story_and_first_segment(story_id: str) -> tuple:
-    """
-    Get story metadata and first segment details.
-
-    Args:
-        story_id: Story UUID
-
-    Returns:
-        Tuple of (story_data, first_segment)
-
-    Raises:
-        ValueError: If story or first segment not found
-        RuntimeError: If database operations fail
-    """
-    # Get story metadata
-    try:
-        story = dynamo.get_item(TableName.STORY, {"StoryID": story_id})
-        if not story:
-            raise ValueError("Story not found")
-    except ClientError as err:
-        logger.error(f"Failed to get story for {story_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to get story: {err}") from err
-
-    # Get first segment
-    first_segment_id = story.get("FirstSegmentID")
-    if not first_segment_id:
-        logger.error(f"Story has no first segment for {story_id}")
-        raise ValueError("Story configuration error")
-
-    first_segment = get_story_segment(story_id, first_segment_id)
-    return story, first_segment
-
-
-def create_active_segment(character_id: str, player_id: str, story_id: str, segment: dict, story_instance_id: str = None) -> dict:
-    """
-    Create an active segment record for tracking progress.
-    Also creates a SegmentHistory record and adds it to StoryHistory if story_instance_id is provided.
-
-    Args:
-        character_id: Character UUID
-        player_id: Player UUID
-        story_id: Story UUID
-        segment: Segment data from Segments table
-        story_instance_id: StoryInstanceID for this story execution (optional for backward compatibility)
-
-    Returns:
-        Active segment record
-
-    Raises:
-        RuntimeError: If database operation fails
-    """
-
-    segment_id = segment.get("SegmentID")
-    segment_type = segment.get("SegmentType", "mechanical")
-    duration = int(segment.get("SegmentDuration", 300))  # Default 5 minutes
-
-    # Use Unix timestamps for storage (required by DynamoDB indexes)
-    start_time = now_unix()
-    end_time = future_unix(duration)
-
-    # Log for debugging with ISO format for readability
-    logger.info(f"Creating segment with StartTime: {from_unix(start_time)}, EndTime: {from_unix(end_time)}, Duration: {duration}s")
-
-    # Generate UUIDv7 for time-based ordering
-    active_segment_id = str(uuid7())
-
-    active_segment = {
-        "ActiveSegmentID": active_segment_id,
-        "CharacterID": character_id,
-        "PlayerID": player_id,
-        "StoryID": story_id,
-        "StoryInstanceID": story_instance_id if story_instance_id else None,  # Store for later use
-        "SegmentID": segment_id,
-        "SegmentType": segment_type,
-        "Status": "active",
-        "StartTime": start_time,
-        "EndTime": end_time,
-        "DefaultStatus": segment.get("ShortStatus", "Processing..."),
-    }
-
-    # Add type-specific fields based on segment type
-    if segment_type == "decision":
-        active_segment["Decision"] = None
-        active_segment["DecisionOptions"] = segment.get("DecisionOptions", {})
-    elif segment_type == "mechanical":
-        # Mechanical segments can have challenges and/or combat
-        active_segment["ChallengeResults"] = []
-        active_segment["Outcome"] = None
-
-        # If combat is configured, set up combat state
-        combat_config = segment.get("Combat", {})
-        if combat_config:
-            # Accept both PascalCase and camelCase for compatibility
-            opponent_id = combat_config.get("OpponentID") or combat_config.get("opponentId")
-
-            # Load opponent to get initial health
-            opponent_health = 5  # Default health
-            if opponent_id:
-                try:
-                    opponent = dynamo.get_item(TableName.OPPONENTS, {"OpponentID": opponent_id})
-                    if opponent:
-                        opponent_health = opponent.get("Health", 5)
-                except Exception as err:
-                    logger.warning(f"Failed to load opponent for combat state init for {opponent_id} Error: {err}")
-
-            # Use PascalCase keys for inter-system records
-            active_segment["CombatState"] = {
-                "Round": 0,
-                "PlayerWounds": [],
-                "OpponentWounds": [],
-                "OpponentHealth": opponent_health,
-                "OpponentID": opponent_id,
-            }
-
-    # Store in DynamoDB
-    try:
-        dynamo.put_item(TableName.ACTIVE_SEGMENTS, active_segment)
-    except ClientError as err:
-        logger.error(f"Failed to create active segment for {active_segment_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to create active segment: {err}") from err
-
-    # Create SegmentHistory record and update StoryHistory if story_instance_id provided
-    if story_instance_id:
-        try:
-            # Create SegmentHistory record
-            segment_history = {
-                "CharacterID": character_id,
-                "ActiveSegmentID": active_segment_id,
-                "StoryInstanceID": story_instance_id,
-                "StoryID": story_id,
-                "SegmentID": segment_id,
-                "SegmentType": segment_type,
-                "StartTime": start_time,
-                "EndTime": end_time,
-                # CharacterUpdates (with XP data) will be added when segment completes
-            }
-            dynamo.put_item(TableName.SEGMENT_HISTORY, segment_history)
-            
-            # Add ActiveSegmentID to StoryHistory's SegmentHistory list
-            dynamo.update_item(
-                TableName.STORY_HISTORY,
-                Key={"CharacterID": character_id, "StoryInstanceID": story_instance_id},
-                UpdateExpression="SET SegmentHistory = list_append(SegmentHistory, :segment_id)",
-                ExpressionAttributeValues={":segment_id": [active_segment_id]},
-            )
-            
-            logger.info(f"Created SegmentHistory record and updated StoryHistory for segment {active_segment_id}")
-        except ClientError as err:
-            # Non-critical - don't fail segment creation if history update fails
-            logger.error(f"Failed to create segment history for {active_segment_id}: {err}")
-
-    return active_segment
-
-
-def create_story_history_entry(character_id: str, story_id: str, story_title: str, story_type: str) -> str:
-    """
-    Create initial history entry for story tracking with new schema.
-
-    Args:
-        character_id: Character UUID
-        story_id: Story UUID
-        story_title: Story title
-        story_type: Type of story (one-time, daily, repeatable)
-
-    Returns:
-        StoryInstanceID (UUIDv7) for this story execution
-
-    Raises:
-        RuntimeError: If database operation fails
-    """
-    try:
-        # Generate UUIDv7 for this story instance
-        story_instance_id = str(uuid7())
-        
-        history_entry = {
-            "CharacterID": character_id,
-            "StoryInstanceID": story_instance_id,
-            "StoryID": story_id,
-            "StoryTitle": story_title,
-            "StartedAt": now_iso(),
-            "StoryType": story_type,
-            "SegmentHistory": [],  # Will contain ActiveSegmentIDs as they're created
-            "SkillXPAwarded": {},
-            "AttributeXPAwarded": {},
-        }
-
-        # Put item (will not overwrite due to unique StoryInstanceID)
-        dynamo.put_item(TableName.STORY_HISTORY, history_entry)
-        
-        logger.info(f"Created story history entry with StoryInstanceID: {story_instance_id}")
-        return story_instance_id
-        
-    except ClientError as err:
-        logger.error(f"Failed to create history entry for {character_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to create history entry: {err}") from err
-
-
-def format_segment_response(segment: dict, active_segment: dict) -> dict:
-    """
-    Format segment data for API response.
-
-    Args:
-        segment: Original segment from Segments table
-        active_segment: Active segment record
-
-    Returns:
-        Formatted response data
-    """
-
-    segment_type = segment.get("SegmentType", "mechanical")
-    time_remaining = max(0, active_segment.get("EndTime", 0) - int(time.time()))
-
-    response = {
-        "SegmentID": active_segment.get("ActiveSegmentID"),
-        "StoryID": active_segment.get("StoryID"),
-        "Type": segment_type,
-        "TimeRemaining": time_remaining,
-    }
-
-    # Add type-specific fields based on documented schema
-    if segment_type == "decision":
-        # DecisionText contains the choice presented
-        response["Content"] = segment.get("DecisionText", "")
-        # Format options from DecisionOptions map
-        decision_options = segment.get("DecisionOptions", {})
-        options = []
-        for option_id, _ in decision_options.items():
-            options.append({"Id": option_id, "Text": option_id.replace("-", " ").title()})  # Format option ID as display text
-        response["Options"] = options
-    elif segment_type == "mechanical":
-        response["ShortStatus"] = segment.get("ShortStatus", "Progressing through the story...")
-        # Include combat opponent if configured
-        combat_config = segment.get("Combat", {})
-        if combat_config:
-            # Accept both PascalCase and camelCase for compatibility
-            opponent_id = combat_config.get("OpponentID") or combat_config.get("opponentId")
-            if opponent_id:
-                response["OpponentID"] = opponent_id
-
-    return response
-
-
-def get_active_decision_segment(character_id: str, player_id: str) -> dict:
-    """
-    Get active decision segment for a character and verify ownership.
-
-    Args:
-        character_id: Character UUID
-        player_id: Cognito user ID for ownership verification
-
-    Returns:
-        Active segment data
-
-    Raises:
-        ValueError: If no active decision segment found or validation fails
-        RuntimeError: If database query fails
-    """
-    try:
-        # Query by CharacterID to find active segment
-        items = dynamo.query(
-            TableName.ACTIVE_SEGMENTS,
-            IndexName="CharacterID-index",
-            KeyConditionExpression="CharacterID = :cid",
-            FilterExpression="PlayerID = :pid AND #status = :status AND SegmentType = :type",
-            ExpressionAttributeNames={"#status": "Status"},
-            ExpressionAttributeValues={
-                ":cid": character_id,
-                ":pid": player_id,
-                ":status": "active",
-                ":type": "decision",
-            },
-        )
-    except ClientError as err:
-        logger.error(f"Failed to query active segments for {character_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to query active segments: {err}") from err
-
-    if not items:
-        logger.warning(f"No active decision segment found for {character_id}")
-        raise ValueError("No active decision segment found")
-
-    active_segment = items[0]
-
-    # Extra validation checks (should already be verified by query)
-    if active_segment.get("PlayerID") != player_id:
-        logger.warning(f"Active segment ownership mismatch for {active_segment.get('ActiveSegmentID')}")
-        raise ValueError("Active segment not found")
-
-    return active_segment
-
-
-def validate_decision_option(active_segment: dict, decision_id: str) -> None:
-    """
-    Validate that the decision is valid for this segment.
-
-    Args:
-        active_segment: Active segment data
-        decision_id: Decision ID submitted by player
-
-    Raises:
-        ValueError: If decision is not valid for this segment
-    """
-    decision_options = active_segment.get("DecisionOptions", {})
-    if decision_id not in decision_options:
-        raise ValueError("Invalid decision option")
-
-    # Check if decision was already made
-    if active_segment.get("Decision"):
-        logger.warning(f"Decision already submitted for {active_segment.get('ActiveSegmentID')}")
-        raise ValueError("Decision already submitted")
-
-
-def update_segment_decision(active_segment_id: str, decision_id: str) -> dict:
-    """
-    Update the active segment with the player's decision.
-
-    Args:
-        active_segment_id: Active segment UUID
-        decision_id: Decision ID chosen by player
-
-    Returns:
-        Updated active segment data
-
-    Raises:
-        RuntimeError: If database update fails
-    """
-    try:
-        # Update the decision field and mark as completed
-        dynamo.update_item(
-            TableName.ACTIVE_SEGMENTS,
-            Key={"ActiveSegmentID": active_segment_id},
-            UpdateExpression="SET #decision = :decision, #status = :status",
-            ExpressionAttributeNames={"#decision": "Decision", "#status": "Status"},
-            ExpressionAttributeValues={":decision": decision_id, ":status": "completed"},
-        )
-
-        # Get updated item
-        updated_segment = dynamo.get_item(TableName.ACTIVE_SEGMENTS, {"ActiveSegmentID": active_segment_id})
-        if not updated_segment:
-            raise RuntimeError("Failed to retrieve updated segment")
-
-        return updated_segment
-    except ClientError as err:
-        logger.error(f"Failed to update active segment for {active_segment_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to update active segment: {err}") from err
-
-
-def get_next_segment_time(active_segment: dict, decision_id: str) -> int:
-    """
-    Calculate the next segment completion time based on the decision.
-
-    Args:
-        active_segment: Active segment data
-        decision_id: Decision ID chosen by player
-
-    Returns:
-        Next segment completion time (0 if no next segment)
-    """
-
-    decision_options = active_segment.get("DecisionOptions", {})
-    next_segment_id = decision_options.get(decision_id)
-
-    if not next_segment_id:
-        return 0
-
-    try:
-        # Get next segment to calculate completion time
-        story_id = active_segment.get("StoryID")
-        if not story_id:
-            return 0
-        next_segment = get_story_segment(story_id, next_segment_id)
-
-        # Next segment will start after processing completes
-        # Add segment duration to get completion time
-        duration = int(next_segment.get("SegmentDuration", 300))
-        return int(time.time()) + duration
-
-    except (ValueError, RuntimeError) as err:
-        logger.error(f"Failed to get next segment for {next_segment_id} Error: {err}")
-        # Continue without next segment time
-        return 0
-
-
-def submit_decision_for_character(character_id: str, decision_id: str, player_id: str) -> dict:
-    """
-    Submit a decision for a character's active decision segment.
-
-    Args:
-        character_id: Character UUID
-        decision_id: Decision ID chosen by player
-        player_id: Authenticated player ID
-
-    Returns:
-        Dict with accepted status and optional next segment time
-
-    Raises:
-        ValueError: If validation fails
-        RuntimeError: If database operations fail
-    """
-
-    # Verify character ownership using player record
-    if not verify_character_ownership(character_id, player_id):
-        raise ValueError("Character not owned by player")
-
-    logger.info(f"Submitting decision for {character_id}")
-
-    # Get active segment for character and verify ownership
-    active_segment = get_active_decision_segment(character_id, player_id)
-    active_segment_id = active_segment.get("ActiveSegmentID")
-    if not active_segment_id:
-        raise ValueError("Active segment ID not found")
-
-    # Validate decision is valid for this segment
-    validate_decision_option(active_segment, decision_id)
-
-    # Update active segment with decision and mark as completed
-    update_segment_decision(active_segment_id, decision_id)
-
-    # Record segment history before advancing
-    story_id = active_segment.get("StoryID")
-    if not story_id:
-        raise ValueError("Story ID not found in active segment")
-
-    # Mark outcome as normal for completed decision segments
-    active_segment["Outcome"] = "normal"
-    active_segment["Decision"] = decision_id
-    active_segment["DecisionMadeAt"] = now_iso()
-
-    record_segment_history(character_id, story_id, active_segment_id, active_segment)
-
-    # Get segment definition to determine next segment
-    story_instance_id = active_segment.get("StoryInstanceID")
-    segment_id = active_segment.get("SegmentID")
-    if not segment_id:
-        raise ValueError("Segment ID not found in active segment")
-
-    segment_def = get_segment_definition(str(story_id), str(segment_id))
-
-    # Determine next segment based on decision
-    next_segment_id = determine_next_segment(segment_def, active_segment, "normal")
-
-    response_data: dict = {
-        "Accepted": True,
-    }
-
-    if next_segment_id:
-        # Create next segment
-        try:
-            next_segment_def = get_segment_definition(story_id, next_segment_id)  # type: ignore
-
-            next_active_segment_id = create_next_active_segment(
-                character_id,
-                player_id,
-                story_id,
-                next_segment_def,
-                story_instance_id,  # Pass instance ID for history tracking
-            )
-
-            # Update character with new active segment
-            update_character_active_segment(character_id, next_active_segment_id)
-
-            # Calculate next segment time using ISO 8601
-            next_segment_duration = next_segment_def.get("SegmentDuration", 60)
-            response_data["NextSegmentTime"] = future_iso(next_segment_duration)
-
-            # Delete prior segment immediately after creating the next to avoid overlap
-            delete_active_segment(active_segment_id)
-
-            logger.info(f"Advanced to next segment after decision and deleted prior segment for {character_id}")
-
-            # Queue mechanical segments for immediate processing
-            if next_segment_def.get("SegmentType") == "mechanical":
-                try:
-                    if SEGMENT_QUEUE_URL:
-                        message_body = {
-                            "ActiveSegmentID": next_active_segment_id,
-                            "CharacterID": character_id,
-                            "StoryID": story_id,
-                            "SegmentID": next_segment_id,
-                            "SegmentType": "mechanical",
-                        }
-                        send_message(SEGMENT_QUEUE_URL, message_body)
-                        logger.info(f"Queued next mechanical segment for processing for {next_active_segment_id}")
-                except Exception as err:
-                    # Non-critical - segment will be picked up by poller
-                    logger.warning(f"Failed to queue mechanical segment for {next_active_segment_id} Error: {err}")
-        except Exception as err:
-            logger.error(f"Failed to create next segment after decision for {next_segment_id} Error: {err}", exc_info=True)
-            raise RuntimeError(f"Failed to create next segment: {err}") from err
-    else:
-        # Story complete
-        complete_story(character_id, story_id, "normal")
-        logger.info(f"Story completed after decision for {character_id}")
-
-    logger.info(f"Decision submitted and story advanced for {active_segment_id}")
-
-    return response_data
-
-
-def format_story_segment_response(active_segment: dict, story_metadata: dict, segment_data: dict) -> dict:
-    """
-    Format story and segment data for API response.
-
-    Args:
-        active_segment: Active segment record from database
-        story_metadata: Story metadata from STORY table
-        segment_data: Segment definition from SEGMENTS table
-
-    Returns:
-        Formatted response dict with story and segment information
-    """
-
-    # Calculate time remaining
-    end_time = int(active_segment.get("EndTime", 0))
-    current_time = int(time.time())
-    time_remaining = max(0, end_time - current_time)
-
-    # Build base response
-    response = {
-        "Story": {
-            "StoryID": active_segment.get("StoryID"),
-            "Title": story_metadata.get("Title", ""),
-            "Type": story_metadata.get("StoryType", ""),
-            "TotalSegments": story_metadata.get("TotalSegments", 1),
-            "CurrentSegmentIndex": segment_data.get("SegmentIndex", 0),
-        },
-        "Segment": {
-            "SegmentID": active_segment.get("SegmentID"),
-            "SegmentType": segment_data.get("SegmentType", ""),
-            "ShortStatus": segment_data.get("ShortStatus", ""),
-            "Description": "",  # Will be set based on segment type
-            "Duration": segment_data.get("SegmentDuration", 0),
-            "TimeRemaining": time_remaining,
-            "StartTime": active_segment.get("StartTime", 0),
-            "EndTime": int(active_segment.get("EndTime", 0)),
-        },
-        "ActiveSegmentID": active_segment.get("ActiveSegmentID", ""),
-        "Status": active_segment.get("Status", ""),
-    }
-
-    # Add segment type specific data
-    segment_type = segment_data.get("SegmentType", "")
-
-    if segment_type == "decision":
-        response["Segment"]["DecisionText"] = segment_data.get("DecisionText", "")
-        # Format options from DecisionOptions map
-        decision_options = segment_data.get("DecisionOptions", {})
-        options = []
-        for option_id, _ in decision_options.items():
-            options.append({"Id": option_id, "Text": option_id.replace("-", " ").title()})
-        response["Segment"]["Options"] = options
-        response["Segment"]["Decision"] = active_segment.get("Decision")
-
-    elif segment_type == "mechanical":
-        # Mechanical segments can contain skill challenges and/or combat
-        response["Segment"]["Description"] = segment_data.get("Description", segment_data.get("Narrative", ""))
-        response["Segment"]["Challenges"] = segment_data.get("Challenges", [])
-        response["Segment"]["ChallengeResults"] = active_segment.get("ChallengeResults", [])
-
-        # Combat is optional within mechanical segments
-        if segment_data.get("Combat"):
-            response["Segment"]["Combat"] = segment_data.get("Combat", {})
-            response["Segment"]["CombatState"] = active_segment.get("CombatState", {})
-
-        response["Segment"]["Outcome"] = active_segment.get("Outcome")
-
-    elif segment_type == "rest":
-        # Rest segments allow wound healing over time
-        response["Segment"]["Description"] = segment_data.get("Description", segment_data.get("Narrative", ""))
-
-    else:
-        # Unknown segment type - add minimal data
-        logger.warning(f"Unknown segment type for {active_segment.get('SegmentID')}")
-        response["Segment"]["Description"] = segment_data.get("Description", segment_data.get("Narrative", ""))
-
-    return response
-
-
-def complete_story_for_character(character_id: str) -> None:
-    """
-    Clean up character state when story is completed.
-
-    Args:
-        character_id: Character UUID
-
-    Raises:
-        RuntimeError: If database operations fail
-    """
-    try:
-        # Update character to clear GameMode and story references
-        dynamo.update_item(
-            TableName.CHARACTERS,
-            Key={"CharacterID": character_id},
-            UpdateExpression="SET GameMode = :none REMOVE ActiveStoryID, ActiveSegmentID",
-            ExpressionAttributeValues={":none": "None"},
-        )
-
-        logger.info(f"Character state cleared for {character_id}")
-    except ClientError as err:
-        logger.error(f"Failed to clear character state for {character_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to clear character state: {err}") from err
-
-
-def complete_story(character_id: str, story_id: str, story_instance_id: str, outcome: str) -> None:
-    """
-    Complete the story, apply rewards, and update character state.
-
-    Args:
-        character_id: Character UUID
-        story_id: Story UUID
-        story_instance_id: Story instance UUID
-        outcome: Final outcome
-    """
-
-    # Clean up character state
-    complete_story_for_character(character_id)
-
-    # Update story history with completion data
-    if story_instance_id:
-        try:
-            # Calculate duration
-            history = dynamo.get_item(TableName.STORY_HISTORY, {"CharacterID": character_id, "StoryInstanceID": story_instance_id})
-            if history:
-                started_at = history.get("StartedAt", "")
-                if started_at:
-                    # Calculate duration in seconds
-                    from datetime import datetime
-                    start_time = datetime.fromisoformat(started_at.replace("Z", "+00:00"))
-                    end_time = datetime.utcnow()
-                    duration = int((end_time - start_time).total_seconds())
-                else:
-                    duration = 0
-                
-                dynamo.update_item(
-                    TableName.STORY_HISTORY,
-                    Key={"CharacterID": character_id, "StoryInstanceID": story_instance_id},
-                    UpdateExpression="SET FinishedAt = :finished, FinalOutcome = :outcome, TotalDuration = :duration",
-                    ExpressionAttributeValues={
-                        ":finished": now_iso(),
-                        ":outcome": outcome,
-                        ":duration": duration,
-                    },
-                )
-        except Exception as err:
-            logger.warning(f"Failed to update story history completion: {err}")
-
-    # Get story metadata for reward calculation and type checking
-    try:
-        story_metadata = dynamo.get_item(TableName.STORY, {"StoryID": story_id})
-        if not story_metadata:
-            raise ValueError("Story not found")
-    except ClientError as err:
-        logger.error(f"Failed to get story for {story_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to get story: {err}") from err
-
-    # Update LastCompleted timestamp for repeatable stories
-    # This is critical for cooldown tracking
-    story_type = story_metadata.get("StoryType", "repeatable")
-    if story_type == "repeatable":
-        try:
-            # Get character's completed stories list
-            character = dynamo.get_item(TableName.CHARACTERS, {"CharacterID": character_id})
-            if character:
-                completed_stories = character.get("CompletedStories", {})
-                completed_stories[story_id] = {
-                    "LastCompleted": int(time.time()),
-                    "CompletionCount": completed_stories.get(story_id, {}).get("CompletionCount", 0) + 1,
-                }
-
-                # Update character with completed story info
-                dynamo.update_item(
-                    TableName.CHARACTERS,
-                    Key={"CharacterID": character_id},
-                    UpdateExpression="SET CompletedStories = :completed",
-                    ExpressionAttributeValues={":completed": completed_stories},
-                )
-                logger.info(f"Updated LastCompleted for repeatable story {story_id} for {character_id}")
-        except Exception as err:
-            # Non-critical - cooldown tracking will just not work properly
-            logger.warning(f"Failed to update LastCompleted for story {story_id}: {err}")
-
-    # Count completed segments from history
-    if story_instance_id:
-        history = dynamo.get_item(TableName.STORY_HISTORY, {"CharacterID": character_id, "StoryInstanceID": story_instance_id})
-        segments_completed = len(history.get("SegmentHistory", [])) if history else 0
-    else:
-        segments_completed = 0
-
-    # Calculate and apply rewards
-    rewards = calculate_story_rewards(story_metadata, outcome, segments_completed)
-    if rewards.get("xp", 0) > 0 or rewards.get("items") or rewards.get("currency", 0) > 0:
-        apply_story_rewards(character_id, rewards)
-
-
-def calculate_story_rewards(story_metadata: dict, outcome: str, segments_completed: int) -> dict:
-    """
-    Calculate rewards based on story outcome and segments completed.
-
-    Args:
-        story_metadata: Story data from STORY table
-        outcome: Final outcome (death, failure, minimal, normal, exceptional)
-        segments_completed: Number of segments completed
-
-    Returns:
-        Dict with calculated rewards (xp, items, etc.)
-    """
-    rewards = {
-        "xp": 0,
-        "items": [],
-        "currency": 0,
-    }
-
-    # No rewards for death
-    if outcome == "death":
-        return rewards
-
-    # Get base XP multiplier and reward tiers
-    base_xp_multiplier = float(story_metadata.get("BaseXPMultiplier", 0.5))
-    reward_tiers_raw = story_metadata.get("RewardTiers", {})
-    # Build a case-insensitive lookup for reward tiers; tolerate non-dict values
-    reward_tiers: dict = {}
-    if isinstance(reward_tiers_raw, dict):
-        reward_tiers = {str(k).lower(): v for k, v in reward_tiers_raw.items()}
-    else:
-        reward_tiers = {}
-
-    # Calculate XP based on outcome
-    outcome_multipliers = {
-        "failure": 0.25,
-        "minimal": 0.5,
-        "normal": 1.0,
-        "exceptional": 1.5,
-    }
-
-    outcome_multiplier = outcome_multipliers.get(outcome, 0)
-    base_xp = story_metadata.get("EstimatedDuration", 300) * base_xp_multiplier
-
-    # XP = base * outcome * (segments_completed / total_segments)
-    total_segments = story_metadata.get("TotalSegments", 1)
-    completion_ratio = min(1.0, segments_completed / max(1, total_segments))
-
-    rewards["xp"] = int(base_xp * outcome_multiplier * completion_ratio)
-
-    # Get items and currency from reward tiers
-    tier_rewards = reward_tiers.get(outcome, {})
-    # If tier rewards are strings (as in some test content), ignore for items/currency
-    if isinstance(tier_rewards, dict):
-        rewards["items"] = tier_rewards.get("items", [])
-        rewards["currency"] = tier_rewards.get("currency", 0)
-    else:
-        rewards["items"] = []
-        rewards["currency"] = 0
-
-    return rewards
-
-
-def apply_story_rewards(character_id: str, rewards: dict) -> None:
-    """
-    Apply calculated rewards to a character.
-
-    Args:
-        character_id: Character UUID
-        rewards: Dict containing xp, items, currency
-
-    Raises:
-        RuntimeError: If database operations fail
-    """
-    try:
-        # Apply XP to character's story skill
-        if rewards.get("xp", 0) > 0:
-            dynamo.update_item(
-                TableName.CHARACTERS,
-                Key={"CharacterID": character_id},
-                UpdateExpression="ADD #skills.#story :xp",
-                ExpressionAttributeNames={
-                    "#skills": "Skills",
-                    "#story": "story",
-                },
-                ExpressionAttributeValues={
-                    ":xp": rewards["xp"],
-                },
-            )
-
-        logger.info(f"Applied story rewards for {character_id}")
-    except ClientError as err:
-        logger.error(f"Failed to apply rewards for {character_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to apply rewards: {err}") from err
-
-
-def apply_combat_rewards(character_id: str, opponent_data: dict) -> None:
-    """
-    Apply rewards from defeating an opponent in combat.
-
-    Args:
-        character_id: Character UUID
-        opponent_data: Opponent data including XPReward and LootTable
-
-    Raises:
-        RuntimeError: If database operations fail
-    """
-    try:
-        # Apply XP reward to combat skill
-        xp_reward = opponent_data.get("XPReward", 10)
-        if xp_reward > 0:
-            dynamo.update_item(
-                TableName.CHARACTERS,
-                Key={"CharacterID": character_id},
-                UpdateExpression="ADD #skills.#combat :xp",
-                ExpressionAttributeNames={
-                    "#skills": "Skills",
-                    "#combat": "combat",
-                },
-                ExpressionAttributeValues={
-                    ":xp": xp_reward,
-                },
-            )
-
-        _ = opponent_data.get("LootTable", [])
-
-        logger.info(f"Applied combat rewards for {character_id}")
-    except ClientError as err:
-        logger.error(f"Failed to apply combat rewards for {character_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to apply combat rewards: {err}") from err
-
-
-def add_segment_to_history(character_id: str, story_instance_id: str, segment_id: str, outcome: str) -> None:
-    """
-    Add a completed segment to the story history.
-
-    Args:
-        character_id: Character UUID
-        story_instance_id: Story instance UUID
-        segment_id: Segment UUID
-        outcome: Segment outcome
-
-    Raises:
-        RuntimeError: If database operations fail
-    """
-    try:
-        segment_entry = {
-            "segmentId": segment_id,
-            "completedAt": now_iso(),
-            "outcome": outcome,
-        }
-
-        dynamo.update_item(
-            TableName.STORY_HISTORY,
-            Key={"CharacterID": character_id, "StoryInstanceID": story_instance_id},
-            UpdateExpression="SET SegmentHistory = list_append(SegmentHistory, :segment)",
-            ExpressionAttributeValues={
-                ":segment": [segment_entry],
-            },
-        )
-
-        logger.info(f"Added segment to history for {character_id}")
-    except ClientError as err:
-        logger.error(f"Failed to add segment to history for {character_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to add segment to history: {err}") from err
-
-
-def update_story_history_xp(character_id: str, story_instance_id: str, skill_xp: dict, attribute_xp: dict) -> None:
-    """
-    Update the story history with accumulated XP from this segment.
-
-    Args:
-        character_id: Character UUID
-        story_instance_id: Story instance UUID
-        skill_xp: Skill XP awarded in this segment
-        attribute_xp: Attribute XP awarded in this segment
-
-    Raises:
-        RuntimeError: If database operation fails
-    """
-    if not skill_xp and not attribute_xp:
-        # No XP to update
-        return
-
-    try:
-        # Build update expressions for XP accumulation
-        update_expressions = []
-        expression_names = {}
-        expression_values = {":zero": Decimal("0")}
-
-        # Update individual fields within the maps
-        # The maps should already exist from ensure_story_history_exists()
-        for skill, xp_value in skill_xp.items():
-            if xp_value > 0:
-                safe_skill = skill.replace("-", "_").replace(" ", "_")
-                # Use placeholder for skill name to handle special characters
-                update_expressions.append(
-                    f"SkillXPAwarded.#skill_{safe_skill} = if_not_exists(SkillXPAwarded.#skill_{safe_skill}, :zero) + :xp_{safe_skill}"
-                )
-                expression_names[f"#skill_{safe_skill}"] = skill
-                expression_values[f":xp_{safe_skill}"] = Decimal(str(xp_value))
-
-        for attribute, xp_value in attribute_xp.items():
-            if xp_value > 0:
-                safe_attr = attribute.replace("-", "_").replace(" ", "_")
-                # Use placeholder for attribute name to handle special characters
-                update_expressions.append(
-                    f"AttributeXPAwarded.#attr_{safe_attr} = if_not_exists(AttributeXPAwarded.#attr_{safe_attr}, :zero) + :xp_attr_{safe_attr}"
-                )
-                expression_names[f"#attr_{safe_attr}"] = attribute
-                expression_values[f":xp_attr_{safe_attr}"] = Decimal(str(xp_value))
-
-        # Add zero value if needed
-        if ":zero" not in expression_values:
-            expression_values[":zero"] = Decimal("0")
-
-        # Execute update
-        update_expression = "SET " + ", ".join(update_expressions)
-
-        dynamo.update_item(
-            TableName.STORY_HISTORY,
-            Key={"CharacterID": character_id, "StoryInstanceID": story_instance_id},
-            UpdateExpression=update_expression,
-            ExpressionAttributeNames=expression_names if expression_names else None,
-            ExpressionAttributeValues=expression_values,
-        )
-
-        logger.info(f"Updated story history with XP for {character_id}")
-    except ClientError as err:
-        logger.error(f"Failed to update story history XP for {character_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to update story history XP: {err}") from err
-
-
-def ensure_story_history_exists(character_id: str, story_id: str, story_title: str) -> None:
-    """
-    Ensure story history record exists.
-
-    Creates a new story history record if one doesn't exist.
-
-    Args:
-        character_id: Character UUID
-        story_id: Story UUID
-        story_title: Story title for display
-
-    Raises:
-        RuntimeError: If database operations fail
-    """
-    try:
-        # Check if story history exists
-        history = dynamo.get_item(
-            TableName.STORY_HISTORY,
-            {"CharacterID": character_id, "StoryID": story_id},
-        )
-
-        if not history:
-            # Create new story history
-            dynamo.put_item(
-                TableName.STORY_HISTORY,
-                {
-                    "CharacterID": character_id,
-                    "StoryID": story_id,
-                    "StoryTitle": story_title,
-                    "StartedAt": now_iso(),
-                    "SkillXPAwarded": {},
-                    "AttributeXPAwarded": {},
-                },
-            )
-            logger.info(f"Created story history record for {character_id}")
-    except ClientError as err:
-        logger.error(f"Failed to ensure story history exists for {character_id} Error: {err}", exc_info=True)
-        raise RuntimeError(f"Failed to ensure story history exists: {err}") from err
